#196 - :recycle: altera o sufixo de arquivos de rotas
